requirements.txt
appdirs==1.4.4
asgiref==3.4.1
astroid==2.4.1
beautifulsoup4==4.9.3
cairocffi==1.2.0
CairoSVG==2.5.2
cffi==1.14.5
cfgv==3.3.0
click==8.0.1
cssselect2==0.4.1
defusedxml==0.7.1
distlib==0.3.2
Django==3.2.5
django-environ==0.4.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
dodgy==0.2.1
filelock==3.0.12
flake8==3.8.4
flake8-polyfill==1.0.2
ghp-import==2.0.1
html5lib==1.1
identify==2.2.10
importlib-metadata==4.6.0
isort==4.3.21
Jinja2==3.0.1
lazy-object-proxy==1.4.3
libsass==0.21.0
Markdown==3.3.4
MarkupSafe==2.0.1
mccabe==0.6.1
mergedeep==1.3.4
mkdocs==1.2.1
mkdocs-material==7.1.9
mkdocs-material-extensions==1.0.1
mkdocs-with-pdf==0.9.3
mkpdfs-mkdocs==1.0.1
nodeenv==1.6.0
packaging==21.0
pep8-naming==0.10.0
Pillow==8.3.0
pre-commit==2.13.0
prospector==1.3.1
psycopg2-binary==2.9.1
pycodestyle==2.6.0
pycparser==2.20
pydocstyle==6.1.1
pyflakes==2.2.0
Pygments==2.9.0
PyJWT==2.1.0
pylint==2.5.3
pylint-celery==0.3
pylint-django==2.1.0
pylint-flask==0.6
pylint-plugin-utils==0.6
pymdown-extensions==8.2
pyparsing==2.4.7
pyphen==0.11.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyyaml-env-tag==0.1
requirements-detector==0.7
setoptconf==0.2.0
six==1.16.0
snowballstemmer==2.1.0
soupsieve==2.2.1
sqlparse==0.4.1
tinycss2==1.1.0
toml==0.10.2
uuid==1.30
virtualenv==20.4.7
watchdog==2.1.3
WeasyPrint==52.5
webencodings==0.5.1
wrapt==1.12.1
zipp==3.5.0
